% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_scatter.R
\name{easy_scatter}
\alias{easy_scatter}
\title{Plot an easy scatter plot from data frame object}
\usage{
easy_scatter(df = datasets::iris, xi = 1, yi = 2, pch = 21, bg = "blue")
}
\arguments{
\item{df}{The data frame object}

\item{xi}{Integer indicating x axis column number}

\item{yi}{Integer indicating y axis column number}

\item{pch}{Default 21}

\item{bg}{Colour name for the points}
}
\value{
Returns a scatter plot
}
\description{
Plot an easy scatter plot from data frame object
}
\examples{
easy_scatter()
easy_scatter(iris,1,2)
easy_scatter(df=iris, xi=1, yi=2,pch=21, bg="blue")
}
