% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_data.R
\docType{data}
\name{student_data}
\alias{student_data}
\title{Generate a student data frame object}
\format{
A data frame with 30 rows and 5 variables by default
\describe{
\item{Name}{as of student name initials from A to M}
\item{Age}{Age of students between 18 to 24 years}
\item{Math}{Mathematics marks from 30 to 90}
\item{Bio}{Biology marks from 30 to 90}
\item{Gender}{Gender as Factor two levels 1 as Male 2 as Female}
}
}
\source{
{Generate as an example data set.}
}
\usage{
student_data
}
\description{
Contains the five variables
}
\examples{
data(student_data) # Lazy loading
}
\keyword{datasets}
