% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_df.R
\name{student_df}
\alias{student_df}
\title{Make a sample data set of students}
\usage{
student_df(n = 30, seed = 999)
}
\arguments{
\item{n}{Integer; Number of rows to generate}

\item{seed}{Integer; for reproducibility}
}
\value{
A data frame with five columns with n rows; Defaults to 30 rows
}
\description{
Make a sample data set of students
}
\examples{
student_df()
student_df(10,888)
student_df(n=10,seed=888)
}
